The local datastore was missing two rather critical constraints that are present in production: 1) Just as you can't perform writes involving entities in multiple entity groups within a single transaction, you also can't perform reads involving entities in multiple entity groups within a single transaction.
2) You can't write to the same entity multiple times within a single
transaction.

Sadly, the ORM was doing both of these things all over the place.  This CL
contains modifications that allow the ORM's testsuite to pass when running
against a local datastore that has these constraints.

The big design change is that 1-to-1 relationships no longer store the key of
the child object in the parent.  There is no way to make this work without
either a getNextId() function in the datastore or support for multiple writes to
a single entity within a transaction.  So, 1-to-1 relationships are implemented
the same way 1-to-many relationships are implemented: the parent has to query
the child table to find its children.  The difference here is that if the query
returns anything other than 0 or 1 results, it's a data error.

Another sizable design change:
We do not support non-owned relationships when running with a transactional
datasource.  This drastically simplifies the logic for determining whether or
not we should auto-assign entity group.  The answer is always yes.

I also updated the distro script to write MANIFEST.MF into the src tree.  Turned
out it was a mistake to have deleted it in an earlier CL.  This way we only need
to maintain the version number in one place.

Let's never do this again.