Don't rely on datastore api to do pojo --> entity type conversions
This fixes a long-standing nasty bug that was the result of our reliance on the
datastore api to do perform the appropriate conversions on non-List Collections,
shorts, ints, bytes, and floats.

The issue was that if a user ever created an Entity and then attempted to fetch
that same Entity, they would get a cached copy, and this cached copy contained
properties of types that the datastore api typically converts (Sets to Lists,
ints to longs, etc).  Our code that constructs pojos from Entities assumes that
the entities it works with will only have Lists, Longs, and Doubles, so when we
pass it one of these Entities we get ClassCastExceptions.  Nasty.