Disable compiled query cache
DataNucleus is caching the compiled form of queries to avoid reparsing.  This is good.  The problem is that, in addition to caching the compiled form of the the queries, it is also caching the symbol table for the queries.  This is bad.  It means that the parameter values are cached along with the query, and when your query is present in the cache you get the old parameter values.  The actual fix involves a change to DataNucleus code and we're trying hard to avoid that before launch.  So, this CL just disables the compiled query cache.